services/Weaviate/requirements.txt
fastapi
uvicorn[standard]
weaviate-client
pydantic
